Fix negative estimated remaining time due to 64-bit integer overflow on large files.
(cherry picked from commit b27d4e7c61b04db054399435b5c57df1f9b1832b)

+++ lib/common/ReadLoggingStream.cpp
@@ -65,7 +65,7 @@ int ReadLoggingStream::Read(void *pBuffer, int NBytes, int Timeout)
 	{
 		box_time_t timeNow = GetCurrentBoxTime();
 		box_time_t elapsed = timeNow - mStartTime;
-		box_time_t finish  = (elapsed * mLength) / mTotalRead;
+		box_time_t finish  = ((float)elapsed * (float)mLength) / mTotalRead;
 		// box_time_t remain  = finish - elapsed;
 		mrLogger.Log(numBytesRead, mOffset, mLength, elapsed, finish);
 	}
